+++ examples/all_data_sources/main.tf
@@ -22,4 +22,8 @@ module "clumio_gcp_connection" {
   clumio_control_plane_id   = clumio_gcp_connection.this.clumio_control_plane_id
   clumio_control_plane_role = clumio_gcp_connection.this.clumio_control_plane_role
   is_gcs_enabled            = var.is_gcs_enabled
+
+  regions                            = var.regions
+  gcs_inventory_bridge_bucket_mode   = var.gcs_inventory_bridge_bucket_mode
+  gcs_inventory_bridge_bucket_labels = var.gcs_inventory_bridge_bucket_labels
 }